REFACTOR watermark & modify modules into pdfconduit.watermark module
