+++ pdfconduit/conduit/__init__.py
@@ -1,4 +1,3 @@
 from pdfconduit.conduit.encrypt import Encrypt
-from pdfconduit.conduit.watermark import Watermark
 
-__all__ = ["Encrypt", "Watermark"]
+__all__ = ["Encrypt"]
