--- pdfconduit/conduit/watermark/__init__.py
@@ -1,3 +1,0 @@
-from pdfconduit.conduit.watermark.watermark import Watermark
-
-__all__ = ["Watermark"]
