+++ pdfconduit/convert/img2pdf.py
@@ -5,8 +5,8 @@ from typing import Optional, List, Union
 
 from PIL import Image
 
-from pdfconduit.modify.canvas import CanvasImg, CanvasObjects
-from pdfconduit.modify.draw import WatermarkDraw
+from pdfconduit.watermark.modify.canvas import CanvasImg, CanvasObjects
+from pdfconduit.watermark.modify.draw import WatermarkDraw
 from pdfconduit.transform.merge import Merge
