--- pdfconduit/modify/canvas/__init__.py
@@ -1,4 +1,0 @@
-from pdfconduit.modify.canvas.constructor import CanvasConstructor
-from pdfconduit.modify.canvas.objects import CanvasImg, CanvasStr, CanvasObjects
-
-__all__ = ["CanvasConstructor", "CanvasImg", "CanvasStr", "CanvasObjects"]
