+++ pdfconduit/watermark/__init__.py
@@ -1,0 +1,4 @@
+from pdfconduit.watermark.watermark import Watermark
+
+
+__all__ = ["Watermark"]
