+++ pdfconduit/watermark/lib/img/watermark.png
