+++ pdfconduit/watermark/modify/canvas/__init__.py
@@ -1,0 +1,4 @@
+from pdfconduit.watermark.modify.canvas.constructor import CanvasConstructor
+from pdfconduit.watermark.modify.canvas.objects import CanvasImg, CanvasStr, CanvasObjects
+
+__all__ = ["CanvasConstructor", "CanvasImg", "CanvasStr", "CanvasObjects"]
