+++ pdfconduit/watermark/modify/draw/__init__.py
@@ -1,5 +1,5 @@
 from PillowImage import PillowImage, img_adjust
 
-from pdfconduit.modify.draw.pdf import DrawPDF, WatermarkDraw
+from pdfconduit.watermark.modify.draw.pdf import DrawPDF, WatermarkDraw
 
 __all__ = ["DrawPDF", "WatermarkDraw", "PillowImage", "img_adjust"]
